refactor(domains): extract pricing into dedicated src/domains/pricing/ subdomain (#822)
Consolidates the pure pricing math under one subdomain so the question
"what does the buyer pay?" has a single owner instead of being
distributed across orders/checkout.ts, promotions/evaluation.ts and
shipping/calculator.ts.

Moves five functions and three types from orders/checkout.ts to
domains/pricing/:
  - calculateOrderPricing, calculateOrderTotals,
    calculateOrderTotalsWithShippingCost, getIncludedTaxAmount,
    OrderPricingLine, roundCurrency  → pricing/order-pricing.ts
  - applyCartDiscounts, CartDiscountInput, CartTotalsBreakdown
    → pricing/cart-totals.ts

orders/checkout.ts keeps the Zod schemas (checkoutFormSchema,
addressSchema, orderItemsSchema) and the SavedCheckoutAddress helpers —
those belong to the checkout flow, not to pricing.

Three call sites updated:
  - createOrder use-case
  - CheckoutPageClient
  - orders-checkout test suite

Charter for pricing/: pure functions only, no DB / Stripe / Prisma /
I/O. Composes primitives owned by orders/, promotions/, shipping/.
Future tax-by-region or order-level fees land here behind clear
function boundaries.

Closes #822.
