+++ src/domains/orders/checkout.ts
@@ -187,89 +187,3 @@ export const orderItemsSchema = z.array(orderItemSchema)
   })
 
 export type OrderItemInput = z.infer<typeof orderItemSchema>
-
-export interface OrderPricingLine {
-  unitPrice: number
-  quantity: number
-  taxRate: number
-}
-
-function roundCurrency(value: number) {
-  return Math.round(value * 100) / 100
-}
-
-export function getIncludedTaxAmount(unitPrice: number, quantity: number, taxRate: number) {
-  const gross = unitPrice * quantity
-  const net = gross / (1 + taxRate)
-  return roundCurrency(gross - net)
-}
-
-export function calculateOrderPricing(lines: OrderPricingLine[]) {
-  const subtotal = roundCurrency(
-    lines.reduce((sum, line) => sum + line.unitPrice * line.quantity, 0)
-  )
-  const taxAmount = roundCurrency(
-    lines.reduce((sum, line) => sum + getIncludedTaxAmount(line.unitPrice, line.quantity, line.taxRate), 0)
-  )
-
-  return {
-    subtotal,
-    taxAmount,
-  }
-}
-
-export function calculateOrderTotals(lines: OrderPricingLine[]) {
-  return calculateOrderTotalsWithShippingCost(lines, 4.95)
-}
-
-export function calculateOrderTotalsWithShippingCost(
-  lines: OrderPricingLine[],
-  shippingCost: number
-) {
-  const { subtotal, taxAmount } = calculateOrderPricing(lines)
-  const grandTotal = roundCurrency(subtotal + shippingCost)
-
-  return {
-    subtotal,
-    taxAmount,
-    shippingCost: roundCurrency(shippingCost),
-    grandTotal,
-  }
-}
-
-export interface CartDiscountInput {
-  subtotalDiscount?: number
-  shippingDiscount?: number
-}
-
-export interface CartTotalsBreakdown {
-  subtotal: number
-  subtotalDiscount: number
-  shippingDiscount: number
-  shipping: number
-  total: number
-}
-
-// Pure composition of the buyer-facing cart totals used by the checkout
-// preview. Lives here (not in CheckoutPageClient.tsx) so the math is
-// covered by unit tests and stays in lock-step with createOrder's
-// server-side totals. Inputs are clamped so a promotion can never
-// produce a negative shipping or grand total.
-export function applyCartDiscounts(
-  baseSubtotal: number,
-  baseShipping: number,
-  { subtotalDiscount = 0, shippingDiscount = 0 }: CartDiscountInput = {},
-): CartTotalsBreakdown {
-  const safeSubtotalDiscount = Math.max(0, subtotalDiscount)
-  const safeShippingDiscount = Math.max(0, shippingDiscount)
-  const shipping = roundCurrency(Math.max(0, baseShipping - safeShippingDiscount))
-  const total = roundCurrency(Math.max(0, baseSubtotal - safeSubtotalDiscount + shipping))
-
-  return {
-    subtotal: roundCurrency(baseSubtotal),
-    subtotalDiscount: roundCurrency(safeSubtotalDiscount),
-    shippingDiscount: roundCurrency(safeShippingDiscount),
-    shipping,
-    total,
-  }
-}
